Renamed the component + Few sparkle improvement
