+++ sparkle/src/components/MessageCard.tsx
@@ -15,7 +15,10 @@ export interface MessageCardProps {
   dismissible?: boolean;
 }
 
-export const MessageCard = React.forwardRef<HTMLDivElement, MessageCardProps>(
+export const MessageCard = React.forwardRef<
+  HTMLDivElement,
+  MessageCardProps
+>(
   (
     {
       className,
