+++ sparkle/src/stories/MessageCard.stories.tsx
